20201110-1706

+++ shared-model/src/main/java/me/dev/entity/Gender.java
@@ -3,7 +3,7 @@ package me.dev.entity;
 import javax.persistence.Entity;
 
 @Entity
-public class Gender extends ImmutableEntity<Integer> {
+public class Gender extends ImmutableEntity<Byte> {
 
     private String code;
     private String title;
